feat(windows): add cpu metrics for docker on windows
