+++ src/biz/metrics_windows.go
@@ -1,6 +1,9 @@
 package biz
 
 import (
+	"math"
+
+	"github.com/docker/docker/api/types"
 	"github.com/newrelic/infra-integrations-sdk/v3/log"
 	"github.com/newrelic/nri-docker/src/raw"
 	gops_mem "github.com/shirou/gopsutil/mem"
@@ -42,4 +45,70 @@ func (mc *MetricsFetcher) memory(mem raw.Memory) Memory {
 	m.UsagePercent = memoryUsagePercent
 
 	return m
+}
+
+// TODO: use container inspect output to get the cpu limit
+func (mc *MetricsFetcher) cpu(metrics raw.Metrics, _ *types.ContainerJSON) CPU {
+	var previous struct {
+		Time int64
+		CPU  raw.CPU
+	}
+	// store current metrics to be the "previous" metrics in the next CPU sampling
+	defer func() {
+		previous.Time = metrics.Time.Unix()
+		previous.CPU = metrics.CPU
+		mc.store.Set(metrics.ContainerID, previous)
+	}()
+
+	cpu := CPU{}
+	cpu.NumProcs = metrics.CPU.NumProcs
+
+	// Reading previous CPU stats
+	if _, err := mc.store.Get(metrics.ContainerID, &previous); err != nil {
+		log.Debug("could not retrieve previous CPU stats for container %v: %v", metrics.ContainerID, err.Error())
+		return cpu
+	}
+
+	// calculate the change for the cpu usage of the container in between readings
+	durationIntervals := uint64(metrics.CPU.Read.Sub(metrics.CPU.PreRead).Nanoseconds()) / 100
+	if durationIntervals <= 0 {
+		log.Debug("duration intervals is 0, returning empty CPU metrics")
+		return cpu
+	}
+
+	maxVal := float64(metrics.CPU.NumProcs * 100)
+
+	cpu.CPUPercent = cpuPercent(previous.CPU, metrics.CPU)
+
+	// If running in windows with Hyper-V isolation, can't check user and kernel CPU usage - set it to 0
+	if metrics.CPU.UsageInUsermode == nil || previous.CPU.UsageInUsermode == nil || metrics.CPU.UsageInKernelmode == nil || previous.CPU.UsageInKernelmode == nil {
+		log.Debug("could not retrieve user and kernel CPU usage for container %v, might be running in Hyper-V isolation", metrics.ContainerID)
+		cpu.UserPercent = 0
+		cpu.KernelPercent = 0
+		return cpu
+	}
+	userDelta := float64(*metrics.CPU.UsageInUsermode - *previous.CPU.UsageInUsermode)
+	cpu.UserPercent = math.Min(maxVal, userDelta*100/float64(durationIntervals))
+
+	kernelDelta := float64(*metrics.CPU.UsageInKernelmode - *previous.CPU.UsageInKernelmode)
+	cpu.KernelPercent = math.Min(maxVal, kernelDelta*100/float64(durationIntervals))
+
+	return cpu
+}
+
+func cpuPercent(previous, current raw.CPU) float64 {
+	// Max number of 100ns intervals between the previous time read and now
+	passIntervals := uint64(current.Read.Sub(current.PreRead).Nanoseconds()) // Start with number of ns intervals
+	passIntervals /= 100                                                     // Convert to number of 100ns intervals
+	passIntervals *= uint64(current.NumProcs)                                // Multiple by the number of processors
+
+	// Intervals used
+	intervalsUsed := current.TotalUsage - previous.TotalUsage
+
+	// Percentage avoiding divide-by-zero
+	if passIntervals > 0 {
+		log.Debug("passIntervals: %v, intervalsUsed: %v", passIntervals, intervalsUsed)
+		return float64(intervalsUsed) / float64(passIntervals) * 100.0
+	}
+	return 0.0
 }
