+++ src/raw/dockerapi/fetcher_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/newrelic/nri-docker/src/constants"
 	"github.com/newrelic/nri-docker/src/raw"
 	"github.com/newrelic/nri-docker/src/raw/dockerapi"
+	"github.com/newrelic/nri-docker/src/util"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
@@ -157,8 +158,8 @@ func Test_Fetch(t *testing.T) {
 	t.Run("CPU metrics", func(t *testing.T) {
 		expectedCPUMetrics := raw.CPU{
 			TotalUsage:        11491000,
-			UsageInKernelmode: 5745000,
-			UsageInUsermode:   5745000,
+			UsageInKernelmode: util.ToPointer(uint64(5745000)),
+			UsageInUsermode:   util.ToPointer(uint64(5745000)),
 			PercpuUsage:       nil,
 			Shares:            2048,
 			ThrottledPeriods:  1,
