+++ test/integration/metrics_test.go
@@ -26,6 +26,7 @@ import (
 	"github.com/newrelic/nri-docker/src/constants"
 	"github.com/newrelic/nri-docker/src/raw"
 	"github.com/newrelic/nri-docker/src/raw/dockerapi"
+	"github.com/newrelic/nri-docker/src/util"
 )
 
 func TestCompareMetrics(t *testing.T) {
@@ -507,8 +508,8 @@ func mockedFileSystem(t *testing.T, hostRoot string) error {
 func inMemoryStorerWithPreviousCPUState() persist.Storer {
 	return inMemoryStorerWithCustomPreviousCPUState(raw.CPU{
 		TotalUsage:        1,
-		UsageInUsermode:   1,
-		UsageInKernelmode: 1,
+		UsageInUsermode:   util.ToPointer(uint64(1)),
+		UsageInKernelmode: util.ToPointer(uint64(1)),
 		PercpuUsage:       nil,
 		ThrottledPeriods:  1,
 		ThrottledTimeNS:   1,
